1. TruckService에 updateRating() 메소드 추가. rating 생성, 수정, 삭제 시 참조하는 트럭의 평점, 리뷰 수 수정을 위함. 2. Truck controller가 개별 트럭을 조회한 후 rating, food 리스트 조회하여 TruckGetResponseDto로 변환하여 반환. 3. Truck 엔티티에 numRating, score 필드 추가.

+++ src/main/java/com/gamakdragons/wheretruck/common/SearchResultDto.java
@@ -12,6 +12,6 @@ import lombok.ToString;
 public class SearchResultDto<T> {
     
     private String status;
-    private long numFound;
+    private int numFound;
     private List<T> docs;
 }
